pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "keryx"
version = "0.1.0"
description = "Portable, human-readable memory for AI agents backed by an Obsidian vault."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "mcp[cli]>=1.2.0,<2.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "pydantic-settings>=2.5.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "typer>=0.12.5,<1.0.0",
  "uvicorn>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
]

[project.scripts]
keryx = "knowledge_gateway.cli:app"
knowledge-gateway = "knowledge_gateway.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_gateway"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
